requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pydantic==2.5.2
python-dotenv==1.0.0
APScheduler==3.10.4
pandas==2.1.3
pytest==7.4.3
